Revert to e98cc813b

+++ framework/Source/GPUImageHueFilter.h
@@ -3,7 +3,7 @@
 
 @interface GPUImageHueFilter : GPUImageFilter
 {
-    GLint hueAdjustUniform;
+    GLfloat hueAdjustUniform;
     
 }
 @property (nonatomic, readwrite) CGFloat hue;
